FIX, REFACRTOE / User Api Response 수정
-[x] Api 별 요구사항에 맞는 Response
-[x] Response 에 맞는 Swagger 수정

Relate to: #54, #11
