+++ src/user/user.service.ts
@@ -12,7 +12,7 @@ import { User } from './entities/user.entity';
 import * as bcrypt from 'bcryptjs';
 import { compare } from 'bcryptjs';
 import { ErrorType } from 'src/common/error.enum';
-import { BossRaidRecord, UserInfoDTO } from './dto/userInfo.dto';
+import { UserInfoDTO } from './dto/userInfo.dto';
 
 @Injectable()
 export class UserService {
@@ -24,7 +24,7 @@ export class UserService {
   /**
    * @작성자 김용민
    * @description 비밀번호 체크, 중복 이메일 확인 후 사용자를 추가합니다.
-  */
+   */
   async createUser(createUserDto: CreateUserDTO): Promise<User> {
     const { email, password, nickname, confirmPassword } = createUserDto;
 
@@ -43,6 +43,7 @@ export class UserService {
 
     try {
       await this.userRepository.save(user);
+
       return user;
     } catch ({ errno, sqlMessage }) {
       if (errno === 1062) {
@@ -60,7 +61,7 @@ export class UserService {
   /**
    * @작성자 김지유
    * @description 유저의 id로 레이드 기록 및 총 점수를 조회합니다.
-  */
+   */
   async getUserInfo(id: number): Promise<UserInfoDTO | undefined> {
     // version 1. 살짝 무식한 방법...
     // createQueryBuilder 디깅 후 Refactoring 예정
@@ -85,7 +86,6 @@ export class UserService {
 
     const userInfo: UserInfoDTO = { totalScore, bossRaidHistory };
 
-    console.log(userInfo);
     // const user = await this.userRepository
     //   .createQueryBuilder('user')
     //   .innerJoinAndSelect('user.raids', 'raids')
@@ -107,7 +107,7 @@ export class UserService {
   /**
    * @작성자 김용민
    * @description 이메일로 사용자를 가져옵니다.
-  */
+   */
   async getUserByEmail(email: string): Promise<User | undefined> {
     const user = await this.userRepository.findOne({ where: { email } });
 
@@ -121,7 +121,7 @@ export class UserService {
   /**
    * @작성자 박신영
    * @description DB에 발급받은 Refresh Token을 암호화하여 저장(bycrypt)
-  */
+   */
   async setCurrentRefreshToken(refreshToken: string, email: string) {
     const salt = await bcrypt.genSalt();
     const hashedRefreshToken = await bcrypt.hash(refreshToken, salt);
@@ -136,7 +136,7 @@ export class UserService {
   /**
    * @작성자 박신영
    * @description 데이터베이스 조회 후 Refresh Token이 유효한지 확인
-  */
+   */
   async getUserRefreshTokenMatches(refreshToken: string, email: string) {
     const user = await this.getUserByEmail(email);
     const isRefreshTokenMatching = await compare(refreshToken, user.hashedRefreshToken);
@@ -149,7 +149,7 @@ export class UserService {
   /**
    * @작성자 박신영
    * @description Refresh Token 값을 null로 바꿈
-  */
+   */
   async removeRefreshToken(id: number) {
     return await this.userRepository.update(id, {
       hashedRefreshToken: null,
@@ -159,7 +159,7 @@ export class UserService {
   /**
    * @작성자 김용민
    * @description id로 사용자를 가져옵니다.
-  */
+   */
   async getUserById(userId: number) {
     const user: User = await this.userRepository.findOne({ where: { id: userId } });
